Add various files to `.gitattibutes` to improve git push/pull performance.
Files added are *.jet and binary lib files, files that are expected to not diff delta compress well.

`.gitattributes` is used to add special handling on a per "file spec" (as in `.gitignore`) basis.  See `man gitattributes` for details.

The `-delta` attribute tells git not to apply diff delta compression to the matched files.

The purpose of this addition is that our n-gram data tends to be 1) very large, and 2) historically not agreeable to diff delta compression.  Because of the two, applying diff delta compression can take a very long time, and ultimately provides virtually zero benefit.  This, in turn, causes git pulls to be quite slow.

Of particular note is that when doing a git pull, when the scren displays `Compressing object...`, git is not doing what you would intuitively think `Compressing` means (i.e., performing `gzip` style compression).  It is in fact computing diff deltas.  This is why tuning git's various zlib compression settings does not result in any performance increases for changes to our n-gram source files.

IMPORTANT: This change is preliminary to see if the expected performance gains match actual performance gains.  This change should be reviewed and appropriate action taken if actual results do not show the expected performance improvemtns.  Also note that the largest performance improvements are expected to be when doing office pulls from the `git.local` cache.